requirements.txt
slack-bolt>=1.18.0
anthropic>=0.23.0
mcp-sdk>=0.1.0
python-dotenv>=1.0.0
